Add Gif Animations and more
A first draft of an improved configuration menu and a new feature for gif animations.

The previous menu had numerous configuration options on one page, which required some scrolling. It has now been organized into sections: General, Sound, and Animations.

Animations is a new feature that displays GIF animations for missed shots, 180, bullseyes, and wins.

Other changes are that you can now have background sound during the game. This feature is not ready at the moment and is under development. The idea is that you should hear background sound from the audience that change the mood depending on what is happening in the game.